Moved DraftEntity to sub namespace.
